WIP: Migrating the definition file into several smaller modules
This will make the definitions easier to read, find, modify, and update.
Adding new modules is as easy as adding a new file without having to
modify a monolithic file.

Progess in coverting everything over to lua_ls and providing better LSP
feedback.
