+++ src/assets/definitions/channelgroups.lua
@@ -1,0 +1,8 @@
+---@meta
+---@diagnostic disable: missing-return, unused-local
+
+---Channel group
+---Used when a Control Script is located inside a Channel Group.  The .Index property is used to determine which Channel Group is currently selected.
+ChannelGroup = {
+	Index = 1,
+}
