+++ src/assets/definitions/ezsvg.lua
@@ -1,0 +1,4 @@
+---@meta
+---@diagnostic disable: missing-return, unused-local
+
+---@class EzSVG
